[ads] SmartNTT code health
