+++ components/brave_ads/core/internal/BUILD.gn
@@ -834,9 +834,6 @@ static_library("internal") {
     "serving/inline_content_ad_serving_feature.h",
     "serving/new_tab_page_ad_serving.cc",
     "serving/new_tab_page_ad_serving.h",
-    "serving/new_tab_page_ad_serving_condition_matcher_util.cc",
-    "serving/new_tab_page_ad_serving_condition_matcher_util_internal.cc",
-    "serving/new_tab_page_ad_serving_condition_matcher_util_internal.h",
     "serving/new_tab_page_ad_serving_delegate.h",
     "serving/new_tab_page_ad_serving_feature.cc",
     "serving/new_tab_page_ad_serving_feature.h",
@@ -941,6 +938,19 @@ static_library("internal") {
     "serving/prediction/model_based/weight/creative_notification_ad_model_based_predictor_weights_builder.cc",
     "serving/prediction/model_based/weight/creative_notification_ad_model_based_predictor_weights_builder.h",
     "serving/prediction/model_based/weight/segment/creative_ad_model_based_predictor_segment_weight_info.h",
+    "serving/targeting/condition_matcher/condition_matcher_util.cc",
+    "serving/targeting/condition_matcher/matchers/epoch_operator_condition_matcher_util.cc",
+    "serving/targeting/condition_matcher/matchers/epoch_operator_condition_matcher_util.h",
+    "serving/targeting/condition_matcher/matchers/internal/epoch_operator_condition_matcher_util_internal.cc",
+    "serving/targeting/condition_matcher/matchers/internal/epoch_operator_condition_matcher_util_internal.h",
+    "serving/targeting/condition_matcher/matchers/pattern_condition_matcher_util.cc",
+    "serving/targeting/condition_matcher/matchers/pattern_condition_matcher_util.h",
+    "serving/targeting/condition_matcher/matchers/regex_condition_matcher_util.cc",
+    "serving/targeting/condition_matcher/matchers/regex_condition_matcher_util.h",
+    "serving/targeting/condition_matcher/prefs/condition_matcher_pref_util.cc",
+    "serving/targeting/condition_matcher/prefs/condition_matcher_pref_util.h",
+    "serving/targeting/condition_matcher/prefs/internal/condition_matcher_pref_util_internal.cc",
+    "serving/targeting/condition_matcher/prefs/internal/condition_matcher_pref_util_internal.h",
     "serving/targeting/segments/top_segments.cc",
     "serving/targeting/segments/top_segments.h",
     "serving/targeting/segments/top_user_model_segments.cc",
