+++ components/brave_ads/core/internal/creatives/new_tab_page_ads/creative_new_tab_page_ads_database_table.cc
@@ -27,7 +27,7 @@
 #include "brave/components/brave_ads/core/internal/segments/segment_util.h"
 #include "brave/components/brave_ads/core/mojom/brave_ads.mojom.h"
 #include "brave/components/brave_ads/core/public/ads_client/ads_client.h"
-#include "brave/components/brave_ads/core/public/serving/new_tab_page_ad_serving_condition_matcher_util.h"
+#include "brave/components/brave_ads/core/public/serving/targeting/condition_matcher/condition_matcher_util.h"
 #include "url/gurl.h"
 
 namespace brave_ads::database::table {
@@ -41,13 +41,12 @@ constexpr char kTableName[] = "creative_new_tab_page_ads";
 
 constexpr int kDefaultBatchSize = 50;
 
-NewTabPageAdConditionMatcherMap StringToConditionMatchers(
-    const std::string& value) {
+ConditionMatcherMap StringToConditionMatchers(const std::string& value) {
   const std::vector<std::string> condition_matchers_as_string =
       base::SplitString(value, ";", base::TRIM_WHITESPACE,
                         base::SPLIT_WANT_NONEMPTY);
 
-  NewTabPageAdConditionMatcherMap condition_matchers;
+  ConditionMatcherMap condition_matchers;
   for (const auto& condition_matcher_as_string : condition_matchers_as_string) {
     const std::vector<std::string> condition_matcher =
         base::SplitString(condition_matcher_as_string, "|",
