+++ components/brave_ads/core/internal/serving/targeting/README.md
@@ -1,5 +1,5 @@
 # Targeting
 
-Targeting helps us place ads in front of users most likely to purchase.
+Targeting helps us place ads in front of users most likely to engage.
 
 Please add to it!
