+++ components/brave_ads/core/internal/serving/targeting/condition_matcher/README.md
@@ -1,0 +1,6 @@
+# Condition Matcher
+
+Condition matchers assess if specific conditions or criteria are fulfilled for
+ad targeting.
+
+Please add to it!
