+++ components/brave_ads/core/internal/serving/targeting/condition_matcher/condition_matcher_util_unittest.cc
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#include "brave/components/brave_ads/core/public/serving/new_tab_page_ad_serving_condition_matcher_util.h"
+#include "brave/components/brave_ads/core/public/serving/targeting/condition_matcher/condition_matcher_util.h"
 
 #include "base/time/time.h"
 #include "brave/components/brave_ads/core/internal/ads_client/ads_client_pref_provider.h"
@@ -15,10 +15,9 @@
 
 namespace brave_ads {
 
-class BraveAdsNewTabPageAdServingConditionMatcherUtilTest
-    : public test::TestBase {
+class BraveAdsConditionMatcherUtilTest : public test::TestBase {
  public:
-  BraveAdsNewTabPageAdServingConditionMatcherUtilTest() : pref_provider_() {
+  BraveAdsConditionMatcherUtilTest() : pref_provider_() {
     // We need to set the clock to 00:00:00 UTC here to ensure the pref registry
     // in `common/test/pref_registry_test_util.cc` is initialized correctly with
     // deterministic default values.
@@ -29,16 +28,15 @@ class BraveAdsNewTabPageAdServingConditionMatcherUtilTest
   const AdsClientPrefProvider pref_provider_;
 };
 
-TEST_F(BraveAdsNewTabPageAdServingConditionMatcherUtilTest,
-       MatchEmptyConditions) {
+TEST_F(BraveAdsConditionMatcherUtilTest, MatchEmptyConditions) {
   // Act & Assert
   EXPECT_TRUE(MatchConditions(&pref_provider_, /*condition_matchers=*/{}));
 }
 
-TEST_F(BraveAdsNewTabPageAdServingConditionMatcherUtilTest,
+TEST_F(BraveAdsConditionMatcherUtilTest,
        MatchConditionsIfAllConditionsAreTrue) {
   // Arrange
-  const NewTabPageAdConditionMatcherMap condition_matchers = {
+  const ConditionMatcherMap condition_matchers = {
       {prefs::kSubdivisionTargetingUserSelectedSubdivision, "AUTO"},
       {prefs::kOptedInToNotificationAds, "1"}};
 
@@ -46,11 +44,9 @@ TEST_F(BraveAdsNewTabPageAdServingConditionMatcherUtilTest,
   EXPECT_TRUE(MatchConditions(&pref_provider_, condition_matchers));
 }
 
-TEST_F(BraveAdsNewTabPageAdServingConditionMatcherUtilTest,
-       MatchEqualOperatorCondition) {
+TEST_F(BraveAdsConditionMatcherUtilTest, MatchEqualOperatorCondition) {
   // Arrange
-  const NewTabPageAdConditionMatcherMap condition_matchers = {
-      {prefs::kServeAdAt, "[=]:7"}};
+  const ConditionMatcherMap condition_matchers = {{prefs::kServeAdAt, "[=]:7"}};
 
   AdvanceClockBy(base::Days(7));
 
@@ -58,11 +54,9 @@ TEST_F(BraveAdsNewTabPageAdServingConditionMatcherUtilTest,
   EXPECT_TRUE(MatchConditions(&pref_provider_, condition_matchers));
 }
 
-TEST_F(BraveAdsNewTabPageAdServingConditionMatcherUtilTest,
-       DoNotMatchEqualOperatorCondition) {
+TEST_F(BraveAdsConditionMatcherUtilTest, DoNotMatchEqualOperatorCondition) {
   // Arrange
-  const NewTabPageAdConditionMatcherMap condition_matchers = {
-      {prefs::kServeAdAt, "[=]:7"}};
+  const ConditionMatcherMap condition_matchers = {{prefs::kServeAdAt, "[=]:7"}};
 
   AdvanceClockBy(base::Days(7) - base::Milliseconds(1));
 
@@ -70,10 +64,9 @@ TEST_F(BraveAdsNewTabPageAdServingConditionMatcherUtilTest,
   EXPECT_FALSE(MatchConditions(&pref_provider_, condition_matchers));
 }
 
-TEST_F(BraveAdsNewTabPageAdServingConditionMatcherUtilTest,
-       MatchPatternCondition) {
+TEST_F(BraveAdsConditionMatcherUtilTest, MatchPatternCondition) {
   // Arrange
-  const NewTabPageAdConditionMatcherMap condition_matchers = {
+  const ConditionMatcherMap condition_matchers = {
       {prefs::kSubdivisionTargetingUserSelectedSubdivision,
        "?UT*"}};  // Value is "AUTO" in the pref.
 
@@ -81,10 +74,9 @@ TEST_F(BraveAdsNewTabPageAdServingConditionMatcherUtilTest,
   EXPECT_TRUE(MatchConditions(&pref_provider_, condition_matchers));
 }
 
-TEST_F(BraveAdsNewTabPageAdServingConditionMatcherUtilTest,
-       DoNotMatchPatternCondition) {
+TEST_F(BraveAdsConditionMatcherUtilTest, DoNotMatchPatternCondition) {
   // Arrange
-  const NewTabPageAdConditionMatcherMap condition_matchers = {
+  const ConditionMatcherMap condition_matchers = {
       {prefs::kSubdivisionTargetingUserSelectedSubdivision,
        "?FOO*"}};  // Value is "AUTO" in the pref.
 
@@ -92,10 +84,9 @@ TEST_F(BraveAdsNewTabPageAdServingConditionMatcherUtilTest,
   EXPECT_FALSE(MatchConditions(&pref_provider_, condition_matchers));
 }
 
-TEST_F(BraveAdsNewTabPageAdServingConditionMatcherUtilTest,
-       MatchRegexCondition) {
+TEST_F(BraveAdsConditionMatcherUtilTest, MatchRegexCondition) {
   // Arrange
-  const NewTabPageAdConditionMatcherMap condition_matchers = {
+  const ConditionMatcherMap condition_matchers = {
       {prefs::kSubdivisionTargetingUserSelectedSubdivision,
        "^AU"}};  // Value is "AUTO" in the pref.
 
@@ -103,10 +94,9 @@ TEST_F(BraveAdsNewTabPageAdServingConditionMatcherUtilTest,
   EXPECT_TRUE(MatchConditions(&pref_provider_, condition_matchers));
 }
 
-TEST_F(BraveAdsNewTabPageAdServingConditionMatcherUtilTest,
-       DoNotMatchRegexCondition) {
+TEST_F(BraveAdsConditionMatcherUtilTest, DoNotMatchRegexCondition) {
   // Arrange
-  const NewTabPageAdConditionMatcherMap condition_matchers = {
+  const ConditionMatcherMap condition_matchers = {
       {prefs::kSubdivisionTargetingUserSelectedSubdivision,
        "^FOO"}};  // Value is "AUTO" in the pref.
 
@@ -114,30 +104,28 @@ TEST_F(BraveAdsNewTabPageAdServingConditionMatcherUtilTest,
   EXPECT_FALSE(MatchConditions(&pref_provider_, condition_matchers));
 }
 
-TEST_F(BraveAdsNewTabPageAdServingConditionMatcherUtilTest,
+TEST_F(BraveAdsConditionMatcherUtilTest,
        DoNotMatchConditionsIfPrefPathWasNotFound) {
   // Arrange
-  const NewTabPageAdConditionMatcherMap condition_matchers = {
-      {"foo.bar", "baz"}};
+  const ConditionMatcherMap condition_matchers = {{"foo.bar", "baz"}};
 
   // Act & Assert
   EXPECT_FALSE(MatchConditions(&pref_provider_, condition_matchers));
 }
 
-TEST_F(BraveAdsNewTabPageAdServingConditionMatcherUtilTest,
+TEST_F(BraveAdsConditionMatcherUtilTest,
        MatchConditionsWithNotOperatorWhenPrefPathNotFound) {
   // Arrange
-  const NewTabPageAdConditionMatcherMap condition_matchers = {
-      {"[!]:foo.bar", "baz"}};
+  const ConditionMatcherMap condition_matchers = {{"[!]:foo.bar", "baz"}};
 
   // Act & Assert
   EXPECT_TRUE(MatchConditions(&pref_provider_, condition_matchers));
 }
 
-TEST_F(BraveAdsNewTabPageAdServingConditionMatcherUtilTest,
+TEST_F(BraveAdsConditionMatcherUtilTest,
        DoNotMatchConditionsIfAllConditionsAreFalse) {
   // Arrange
-  const NewTabPageAdConditionMatcherMap condition_matchers = {
+  const ConditionMatcherMap condition_matchers = {
       {prefs::kOptedInToNotificationAds, "0"},  // Value is "1" in the pref.
       {prefs::kServeAdAt, "[>]:7"}};            // 5 days ago in the pref.
