+++ components/brave_ads/core/internal/serving/targeting/condition_matcher/matchers/README.md
@@ -1,0 +1,5 @@
+# Condition Matchers
+
+Condition matchers evaluate whether specific criteria for ad targeting are met by comparing them against operators and patterns.
+
+Please add to it!
