+++ components/brave_ads/core/internal/serving/targeting/condition_matcher/matchers/regex_condition_matcher_util.h
@@ -1,0 +1,18 @@
+/* Copyright (c) 2024 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+#ifndef BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_SERVING_TARGETING_CONDITION_MATCHER_MATCHERS_REGEX_CONDITION_MATCHER_UTIL_H_
+#define BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_SERVING_TARGETING_CONDITION_MATCHER_MATCHERS_REGEX_CONDITION_MATCHER_UTIL_H_
+
+#include <string_view>
+
+namespace brave_ads {
+
+// Matches a value against a regular expression condition.
+bool MatchRegex(std::string_view value, std::string_view condition);
+
+}  // namespace brave_ads
+
+#endif  // BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_SERVING_TARGETING_CONDITION_MATCHER_MATCHERS_REGEX_CONDITION_MATCHER_UTIL_H_
