+++ components/brave_ads/core/internal/serving/targeting/condition_matcher/prefs/README.md
@@ -1,0 +1,5 @@
+# Condition Matcher Prefs
+
+Condition matchers are utilized to assess if specific conditions or criteria are fulfilled for ad targeting by matching against profile, local state, and virtual preferences.
+
+Please add to it!
