+++ components/brave_ads/core/public/BUILD.gn
@@ -47,7 +47,7 @@ source_set("headers") {
     "prefs/pref_provider.h",
     "prefs/pref_provider_interface.h",
     "prefs/pref_registry.h",
-    "serving/new_tab_page_ad_serving_condition_matcher_util.h",
+    "serving/targeting/condition_matcher/condition_matcher_util.h",
     "targeting/geographical/subdivision/supported_subdivisions.h",
     "user_attention/user_idle_detection/user_idle_detection_feature.h",
     "user_engagement/ad_events/ad_event_cache.h",
