+++ components/brave_ads/core/test/BUILD.gn
@@ -463,8 +463,6 @@ source_set("brave_ads_unit_tests") {
     "//brave/components/brave_ads/core/internal/serving/inline_content_ad_serving_delegate_mock.h",
     "//brave/components/brave_ads/core/internal/serving/inline_content_ad_serving_feature_unittest.cc",
     "//brave/components/brave_ads/core/internal/serving/inline_content_ad_serving_unittest.cc",
-    "//brave/components/brave_ads/core/internal/serving/new_tab_page_ad_serving_condition_matcher_util_internal_unittest.cc",
-    "//brave/components/brave_ads/core/internal/serving/new_tab_page_ad_serving_condition_matcher_util_unittest.cc",
     "//brave/components/brave_ads/core/internal/serving/new_tab_page_ad_serving_delegate_mock.cc",
     "//brave/components/brave_ads/core/internal/serving/new_tab_page_ad_serving_delegate_mock.h",
     "//brave/components/brave_ads/core/internal/serving/new_tab_page_ad_serving_feature_unittest.cc",
@@ -526,6 +524,13 @@ source_set("brave_ads_unit_tests") {
     "//brave/components/brave_ads/core/internal/serving/prediction/model_based/weight/creative_notification_ad_model_based_predictor_weights_builder_unittest.cc",
     "//brave/components/brave_ads/core/internal/serving/prediction/model_based/weight/segment/creative_ad_model_based_predictor_segment_weight_test_util.cc",
     "//brave/components/brave_ads/core/internal/serving/prediction/model_based/weight/segment/creative_ad_model_based_predictor_segment_weight_test_util.h",
+    "//brave/components/brave_ads/core/internal/serving/targeting/condition_matcher/condition_matcher_util_unittest.cc",
+    "//brave/components/brave_ads/core/internal/serving/targeting/condition_matcher/matchers/epoch_operator_condition_matcher_util_unittest.cc",
+    "//brave/components/brave_ads/core/internal/serving/targeting/condition_matcher/matchers/internal/epoch_operator_condition_matcher_util_internal_unittest.cc",
+    "//brave/components/brave_ads/core/internal/serving/targeting/condition_matcher/matchers/pattern_condition_matcher_util_unittest.cc",
+    "//brave/components/brave_ads/core/internal/serving/targeting/condition_matcher/matchers/regex_condition_matcher_util_unittest.cc",
+    "//brave/components/brave_ads/core/internal/serving/targeting/condition_matcher/prefs/condition_matcher_pref_util_unittest.cc",
+    "//brave/components/brave_ads/core/internal/serving/targeting/condition_matcher/prefs/internal/condition_matcher_pref_util_internal_unittest.cc",
     "//brave/components/brave_ads/core/internal/serving/targeting/segments/top_segments_unittest.cc",
     "//brave/components/brave_ads/core/internal/serving/targeting/segments/top_user_model_segments_unittest.cc",
     "//brave/components/brave_ads/core/internal/serving/targeting/segments/top_user_model_segments_util_unittest.cc",
